requirements.txt
mysql-connector-python==8.0.32
tqdm==4.65.0
python-dotenv==1.0.0
pymediainfo==6.0.1
flask==2.3.2
flask-wtf==1.1.1
flask-cors==3.0.10
flask-sqlalchemy==3.0.5
apscheduler==3.10.1
werkzeug==2.3.6
pandas==2.0.3
matplotlib==3.7.2
sqlalchemy==2.0.15
flask-apscheduler==1.12.3
